Convert methodId to enum

+++ src/RabbitMQ.Next.Transport/Methods/Channel/CloseMethodFormatter.cs
@@ -8,6 +8,6 @@ namespace RabbitMQ.Next.Transport.Methods.Channel
         public Span<byte> Write(Span<byte> destination, CloseMethod method) =>
             destination.Write((ushort) method.StatusCode)
                 .Write(method.Description)
-                .Write(method.FailedMethodId);
+                .Write((uint)method.FailedMethodId);
     }
 }
