+++ tests/RabbitMQ.Next.Tests/TopologyBuilder/ExchangeBindingBuilderTests.cs
@@ -57,7 +57,7 @@ namespace RabbitMQ.Next.Tests.TopologyBuilder
         {
             var channel = Substitute.For<IChannel>();
             channel.SendAsync<BindMethod, BindOkMethod>(default)
-                .ReturnsForAnyArgs(Task.FromException<BindOkMethod>(new ChannelException((ushort)replyCode, "error message", (uint) MethodId.ExchangeBind)));
+                .ReturnsForAnyArgs(Task.FromException<BindOkMethod>(new ChannelException((ushort)replyCode, "error message", MethodId.ExchangeBind)));
             var builder = new ExchangeBindingBuilder("destination", "source");
 
             await Assert.ThrowsAsync(exceptionType,async ()=> await builder.ApplyAsync(channel));
