+++ tests/RabbitMQ.Next.Tests/TopologyBuilder/QueueBuilderTests.cs
@@ -60,7 +60,7 @@ namespace RabbitMQ.Next.Tests.TopologyBuilder
         {
             var channel = Substitute.For<IChannel>();
             channel.SendAsync<DeclareMethod, DeclareOkMethod>(default)
-                .ReturnsForAnyArgs(Task.FromException<DeclareOkMethod>(new ChannelException((ushort)replyCode, "error message", (uint) MethodId.QueueBind)));
+                .ReturnsForAnyArgs(Task.FromException<DeclareOkMethod>(new ChannelException((ushort)replyCode, "error message", MethodId.QueueBind)));
             var builder = new QueueBuilder("queue");
 
             await Assert.ThrowsAsync(exceptionType,async ()=> await builder.ApplyAsync(channel));
